requirements.txt
fastapi==0.115.0
Pillow==10.4.0
pydantic==2.9.2
uvicorn==0.31.0